Add support for SystemVerilog sign cast
SystemVerilog supports sign cast where it is possible to change the
signedness an expression. Syntactical it is similar to width or type
casting, except that the keywords 'signed' or 'unsigned' are used in front
of the cast operator.

E.g.

  logic [3:0] a = 4'b1000;
  logic [7:0] b = signed'(a); // b is 8'b11111000;
  logic signed [3:0] c = 4'b1000;
  logic signed [7:0] d = unsigned'(c); // d is 8'b00001000;

As noted by the LRM section 6.24.1 ("Cast operator") applying a signing
cast to an expression is equivalent as calling the $signed() and $unsigned()
system tasks on the expression.

To keep things simple implement it as just that and instantiate a call to
the corresponding system task when encountering a sign cast.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>