Add support for part select on concatenations
SystemVerilog adds support for part selects on concatenation
expressions. For the part select the concatenation expression
looks as if the part select was done on integral identifier [N-1:0]
where N is the width of the concatenation.

Part select on concatenations is especially useful since it
allows part selects on arbitrary expressions by wrapping
them in curly braces. E.g. `{a+b}[10:1]`.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>