Correctly report struct member type
The base type of a struct member can be different from the
base type of the struct.

E.g. if a struct contains a mix of 2-state and 4-state members
the struct itself will be 4-state, but the 2-state members still
will be 2 state and need to be cast appropriately when accessed.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>