tgt-vvp: Handle type cast when evaluating select
A part select expression can, in addition to selecting part
of a net, do a type cast. For example accessing a 2-state member
of a 4-state packed struct requires a cast.

Insert a cast operation when evaluating the select expression
if necessary.

Note that part selects are only defined for the vector types
and cast from 2-state to 4-state does not require any special handling.
So only cast from 4-state to 2-state needs to be handled.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>