tgt-vvp: Handle out-of-bounds negative indices in VPI references
The VVP runtime does not support negative immediate indices in VPI
references.

Make sure to fallback to on-stack evaluation of the value when a part
select with negative out-of-bound index is passed to a VPI call.

E.g. the following code will elaborate fine, but produce a syntax error
when running through vvp.

```
logic [1:0] a;
$display(a[1:-1]);
```

Since this is a rather unlikely corner case there is no need to extend the
VVP runtime itself to handle it.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>