tgt-vvp: Cast reads from 2-state dynamic arrays/queues
vvp does not track whether the values stored in a
dynamic array or queue are 2-state or 4-state. Internally the data
is always stored as 4-state.

To make sure that the read value is actually 2-state do a cast
for 2-state reads.

E.g. performing an out-of-bounds access on a 2-state
dynamic array or queue will yield a undefined value.


Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>