tgt-vvp: Cast part select to 2-state if required
A part select can yield an undefined value if the indexed
part is out-of-bounds.

In addition the part select expression type can also be different
from the type of the sub-expression. This is e.g. the case
when accessing a 2-state struct member in a struct with mixed state
members.

In both cases a cast is required for correct behavior.

The cast can only be skipped if the source expression is 2-state and
the part select is constant and fully within the bounds of the
sub-expression.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>