Add a typedef for std::vector<netrange_t>
std::vector<netrange_t> is used for signal array dimensions. As such it is
used in quite a few places.

It is a bit of a 'mouthful' to write. Add a typedef that can be used as a
shorthand to refer to it. This helps to keep lines where this is used from
growing to overly long.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>