NetNet: Pass unpacked dimensions as std::vector instead of std::list
Most places in the code use a std::vector for array dimensions.
The only exception is the constructor of NetNet, which uses
a std::list to pass the unpacked dimensions. But to store the
unpacked dimensions it also uses a std::vector.

There does not seem to be a good reason why the constructor
has to take a std::list, so switch it also to std::vector.

This allows to simplify the code and remove some special handling
for std::list<netrange_t>.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>