Elaborate unpacked array concatenation values with the correct context
Values in a unpacked array concatenation are evaluated in a self-determined
context, but the result must than be width cast to element type of the
array.

E.g. the following two are equivalent:

```
int d[]= {a, b};
```

```
int d[] = new [2];
d[0] = {a};
d[1] = {b};
```

and should give the same results.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>