Elaborate unpacked array concatenation values with the correct context
Values in a unpacked array concatenation are evaluated in a self-determined
context, but the result must then be assignment cast to element type of the
array.

This is very similar how values in a non-array concatenation are evaluated.

E.g. the following two are equivalent

```
int unsigned d[]= {a, b};
```

```
int unsigned d[] = new [2];
d[0] = {a};
d[1] = {b};
```

and should give the same results.

But there are some important differences.

1) Non-array concatenation only works on vectorable expressions, while
   array concatenation works on any expression with a self-determined
   type.
2) Non-array concatenation will evaluate the expression always as unsigned.
   While for array expression the signedness of the expression is preserved.
   This means that signed values will be sign extended if their width is
   smaller than the array element type.

This is different from array pattern assignment where the expressions are
evaluated in a context of the array element type.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>