ivtest: br_ml20150606: Change invalid port declaration
The br_ml20150606 regression test uses ANSI style port declarations, but
also re-declares the ports inside module as signals.

This is not valid (System)Verilog, even though the data type in the port
declaration is an implicit type. An ANSI-style port is always fully
defined. This defined section ... of the LRM (1800-2017).

Currently the test passes because this invalid declaration is currently not
detected. But the code will be updated to detect and report an error for
this.

In preparation for this update the regression test to use non-ANSI style
port declarations.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>