Add support for forward type declarations
SystemVerilog supports forward type declarations. This allows to declare a
that a identifier is a type and use that type identifier, e.g. in a signal
declaration, before declaring what the actual type is. The type still needs
to be fully defined eventually in the same scope as its forward type
declaration. E.g.

```
typedef T;
T x;
typedef int T;
```

The forward type definition can also contain the kind of the type its is
going to be. E.g struct, union, class, etc. The LRM calls this the basic
type. If the actual type is not of the basic type specified in the
forward type declaration this is an error. E.g.

```
typedef struct T;
typedef int T; // Error, int is not a struct
```

It is legal to more than one forward type declaration for the same type
name, as long as the basic type is the compatible. It is even legal to have
a forward type declaration after the actual type has already been declared.
E.g.

```
typedef T;
typedef int T;
typedef T;
```

Implement support for forward type definitions as part of the new
typedef_t. The basic type will be attached to the typedef_t.

The compatibility of the basic type for multiple forward type declarations
will be checked in the parser. The compatibility of the basic type to the
actual type will be checked during elaboration, once the actual type is
known.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>