tgt-vlog95: Consider scopes with only localparams as unique
For module with parameters the vlog95 backend generates one module
declaration for each module instance. This is so that different values for
the module parameters can be supported.

Local parameters are guaranteed to have the same value for all module
instances though. Add support for detecting the case that all module
parameters are local parameters and in that case only create one shared
module declaration.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>