Allow to override type identifiers with identifiers in nested scopes
In SystemVerilog it is possible to override an existing identifier or type
identifier from an outer scope. It is possible to override a type
identifier with an identifier and vice-versa. E.g. the following is legal

```
typedef int T;
int X;

module test;
  typedef int X;
  int T;
endmodule
```

In the unit scope T is a type and X is a variable, whereas in the module
scope X is a type and T is a variable.

Currently it is possible to override a identifier from an outer scope with
a type identifier. But it is not possible to override a type identifier
with an identifier.

Adding support for this in the parser is non-trivial

E.g. `input T` could either be finished as `input T;` in which case T is an
identifier or `input T x;` in which case T is a type. Having a naive
rule `input <data_type_or_implicit> <identifier_or_type_identifier>` will
result in a shift/reduce conflict.

What complicates this even more is that type identifiers can have packed
dimensions attached to it, which looks the same as unpacked dimensions on
an identifier. E.g. `input T [1:0]` could be finished as `input T [1:0];`
or `input T [1:0] x;`.

To resolve this conflict all possible combinations of type, identifier and
dimensions need to be formulated as single rule. This way the parser can't
reduce until the declaration is fully parsed.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>