vvp: Handle non-representable indices for load and store
When a vector is loaded into an index register flag 4 is set
 * 1 if the value is undefined
 * x if the value is defined, but not representable as a 64-bit integer
 * 0 otherwise

Currently most load or store instructions that use an index register will
skip the load or store only if the index register is 1.

This means that if the index vector value is not representable as 64-bit
integer the load or store will still happen, but with the wrapped around
value.

E.g.
```
reg [1:0] x = 2'b00;
reg [64:0] i = 2**64;
x[i] = 1'b1;
$display(x); // will print 2'b01
```

Make sure to only perform the load or store if flag 4 is 0 and skip it
otherwise. This makes sure that if the index value is not representable as
a 64-bit integer the load or store will also be skipped.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>