Remove pre C++11 unique_ptr compat code
For compatibility with version of C++ earlier than 11 there is a
compatibility layer that defines unique_ptr as auto_ptr. And while the two
are similar there are subtle differences that can lead to bugs. Defining
unique_ptr as auto_ptr only works if the scope how it is used it carefully
restricted.

Now that the project requires C++11 remove the compatibility layer.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>