Add support for assignment pattern repeat
SystemVerilog allows repetition for assignment patterns. This works similar
to repetition in concatenations and also uses a similar syntax.

The repetition is specified as a constant expression followed by a list of
expressions in curly brackets. Each expression is then only evaluated once,
but the resulting values are inserted multiple times into the target array.

E.g. `'{2{1, 2}}` is equivalent to `'{1, 2, 1, 2}`.

Add support for handling repetitions for assignment patterns. This requires
changes at all levels, the parser, the elaboration as well as the code
generation backend.

The implementation is done similar to concatenation repetition
implementation and shares some code with it.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>