Add support for package scoped task calls
It is possible to have a statement that invokes a packaged scoped task or
void function. E>g.
```
initial P::t;
```

Currently there is no rule in the parser for handling this.
The naive way of implementing this leads to a shift/reduce conflict in the
parser. This is because something like `P::X` at the beginning of a block
could either complete as a variable declaration of the package scoped type
X, e.g. `P::X x;`, or it could complete as a task call like `P::X;`

What cause the conflict is that there are optional elements that can appear
either before the data declaration of the task call. These optional items
take the form of

```
X_opt
 : X
 |
 ;
...
Z
 : X_opt Y ...
 ...
 ;
```

When X is not found before Y, X_opt gets reduced using the empty rule. But
now that X has been reduced it can't be undone.

To resolve this transform the grammar for the affected items to use the
following schema instead.

```
X_opt_plus_Y
 : X Y
 | Y
 ...
 ;
Z
 : X_opt_plus_Y ...
 ...
 ;
```

This way there are no empty reductions and no conflicts. This comes at the
expense of one more item on the parser stack, but that should be OK.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>