Allow class properties to be shadowed by local symbols
Class properties can be shadowed by local symbols in a class method.
E.g.

Name resolution currently tries to elaborate a symbol as a class property
before even searching for the symbol. As a result the class property will
be used instead of the local symbol shadowing it. E.g.

```
class C;
  int x = 1;
  task t();
    int x = 2;
    $display(x); // Should print 2, will print 1
  endtask
endclass
```

The class property elaboration also ignores the package scope of the
identifier resulting in access to a class property being elaborated if
there is a property of the same name as the scoped identifier. E.g.

```
package P;
  int x = 2;
endpackage

class C;
  int x = 1;
  task t;
    $display(P::x); // Should print 2, will print 1
  endtask
endclass
```

To fix both of these issues first run the symbol search and only elaborate
it as a class property if symbol search reports that the identifier is a
class property.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>