Fix $left/$right for C style unpacked arrays
Unpacked array dimensions that are implemented with only a single size
value have a $left of 0 and a $right of size - 1. E.g. `x[10]` is
equivalent to `x[0:9]`.

Currently it is implemented the other way around. There are a few contexts
where this distinction matters. For example array to array assignments,
which are done left-to-right.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>